+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Deng 2024 magnesium sulfate ([doi:10.1186/s12884-024-06620-x](https://doi.org/10.1186/s12884-024-06620-x)) -- Chinese women with preeclampsia.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.